Moved disaggregate outside of the ContextMaker
